+++ examples/sample_benchmarks/using_options.cpp
@@ -8,8 +8,8 @@
 
 #include <gauge/gauge.hpp>
 
-/// This test shows how you can add options to the benchmark
-/// options
+// This test shows how you can add options to the benchmark
+// options
 class using_options : public gauge::time_benchmark
 {
 public:
@@ -23,13 +23,8 @@ public:
         uint32_t size = cs.get_value<uint32_t>("array_size");
 
         m_vector.resize(size);
-        for(uint32_t i = 0; i < size; ++i)
-            m_vector[i] = rand() % 256;
-    }
 
-    void tear_down()
-    {
-        // No implementation required
+        std::generate(m_vector.begin(), m_vector.end(), rand);
     }
 
     void get_options(gauge::po::variables_map& options)
@@ -38,7 +33,7 @@ public:
         // get_options() function and using the option name and
         // type to access the variables_map.
         auto values = options["array_size"].as<std::vector<int> >();
-        for(auto v : values)
+        for (auto v : values)
         {
             gauge::config_set cs;
             cs.set_value<uint32_t>("array_size", v);
@@ -50,40 +45,14 @@ public:
 
 };
 
-// Counting bits set, Brian Kernighan's way
-uint32_t count_bk(uint32_t v)
-{
-    uint32_t c; // c accumulates the total bits set in v
-
-    for (c = 0; v; c++)
-    {
-        v &= v - 1; // clear the least significant bit set
-    }
-    return c;
-}
-
-// Counting bits set (naive way)
-uint32_t count_naive(uint32_t v)
-{
-    uint32_t c; // c accumulates the total bits set in v
-
-    for (c = 0; v; v >>= 1)
-    {
-        c += v & 1;
-    }
-    return c;
-}
-
-/// Using this macro we may specify options. For specifying options
-/// we use the boost program options library. So you may additional
-/// details on how to do it in the manual for that library.
+// Using this macro we may specify options. For specifying options
+// we use the boost program options library. So you may additional
+// details on how to do it in the manual for that library.
 BENCHMARK_OPTION(array_size)
 {
     gauge::po::options_description options;
 
-    std::vector<int> size;
-    size.push_back(10);
-    size.push_back(40);
+    std::vector<int> size = {10U, 40U};
 
     auto default_size =
         gauge::po::value<std::vector<int> >()->default_value(
@@ -95,24 +64,52 @@ BENCHMARK_OPTION(array_size)
     gauge::runner::instance().register_options(options);
 }
 
-BENCHMARK_F(using_options, CountBits, count_bk, 10)
+// Counting bits set, Brian Kernighan's way
+inline uint32_t count_bk(uint32_t v)
 {
+    uint32_t c; // c accumulates the total bits set in v
 
-    volatile uint32_t sum;
-    RUN{
-        for(auto& v: m_vector)
-            sum = count_bk(v);
+    for (c = 0; v; c++)
+    {
+        v &= v - 1; // clear the least significant bit set
     }
-    (void)sum; // Suppress warning about unused variable
+    return c;
 }
 
-BENCHMARK_F(using_options, CountBits, count_naive, 10)
+// Counting bits set (naive way)
+inline uint32_t count_naive(uint32_t v)
 {
+    uint32_t c; // c accumulates the total bits set in v
 
-    volatile uint32_t sum;
-    RUN{
-        for(auto& v: m_vector)
-            sum = count_naive(v);
+    for (c = 0; v; v >>= 1)
+    {
+        c += v & 1;
     }
-    (void)sum; // Suppress warning about unused variable
+    return c;
+}
+
+BENCHMARK_F_INLINE(using_options, CountBits, count_bk, 10)
+{
+    volatile uint32_t sum;
+
+    RUN
+    {
+        for(auto v: m_vector)
+            sum += count_bk(v);
+    }
+
+    (void) sum; // Suppress warning about unused variable
+}
+
+BENCHMARK_F_INLINE(using_options, CountBits, count_naive, 10)
+{
+    volatile uint32_t sum;
+
+    RUN
+    {
+        for(auto v: m_vector)
+            sum += count_naive(v);
+    }
+
+    (void) sum; // Suppress warning about unused variable
 }