+++ test/src/test_time.cpp
@@ -14,12 +14,6 @@ namespace bc = boost::chrono;
 
 struct sleep_benchmark : public gauge::time_benchmark
 {
-    bool accept_measurement()
-    {
-        // We always accept the first measurement here
-        return true;
-    }
-
     double measurement()
     {
         // Get the time spent sleeping
@@ -43,27 +37,27 @@ protected:
     uint32_t m_last_delay;
 };
 
-BENCHMARK_F(sleep_benchmark, Gauge, sleep_10usecs, 1)
+BENCHMARK_F_INLINE(sleep_benchmark, Gauge, sleep_10usecs, 1)
 {
      run_benchmark(10);
 }
 
-BENCHMARK_F(sleep_benchmark, Gauge, sleep_100usecs, 1)
+BENCHMARK_F_INLINE(sleep_benchmark, Gauge, sleep_100usecs, 1)
 {
     run_benchmark(100);
 }
 
-BENCHMARK_F(sleep_benchmark, Gauge, sleep_1000usecs, 1)
+BENCHMARK_F_INLINE(sleep_benchmark, Gauge, sleep_1000usecs, 1)
 {
     run_benchmark(1000);
 }
 
-BENCHMARK_F(sleep_benchmark, Gauge, sleep_10000usecs, 1)
+BENCHMARK_F_INLINE(sleep_benchmark, Gauge, sleep_10000usecs, 1)
 {
     run_benchmark(10000);
 }
 
-BENCHMARK_F(sleep_benchmark, Gauge, sleep_100000usecs, 1)
+BENCHMARK_F_INLINE(sleep_benchmark, Gauge, sleep_100000usecs, 1)
 {
     run_benchmark(100000);
 }
@@ -77,3 +71,22 @@ TEST(Gauge, sleep_intervals)
     gauge::runner::add_default_printers();
     gauge::runner::run_benchmarks(argc, argv);
 }
+
+
+// TEST(Gauge, reference_sleep_intervals)
+// {
+//     uint32_t loop = 100000;
+
+//     auto start = bc::high_resolution_clock::now();
+
+//     for(uint32_t i = 0; i < loop; ++i)
+//         boost::this_thread::sleep_for(bc::microseconds(10));
+
+//     auto stop = bc::high_resolution_clock::now();
+
+//     auto duration = static_cast<double>(
+//         bc::duration_cast<bc::microseconds>(stop - start).count());
+
+//     std::cout << "Duration " << duration << std::endl;
+//     std::cout << "Duration per loop" << duration/loop << std::endl;
+// }